# src/main/resources/functional_requirements/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/ui/party/{id}": {
      "get": {
        "tags": [
          "party-controller"
        ],
        "operationId": "getPartyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataParty"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "party-controller"
        ],
        "operationId": "updateParty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transition",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Party"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataParty"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "party-controller"
        ],
        "operationId": "deleteParty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/ui/party": {
      "get": {
        "tags": [
          "party-controller"
        ],
        "operationId": "getAllParties",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataParty"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "party-controller"
        ],
        "operationId": "createParty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Party"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataParty"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui/party/business/{partyId}": {
      "get": {
        "tags": [
          "party-controller"
        ],
        "operationId": "getPartyByBusinessId",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityWithMetadataParty"
                }
              }
            }
          }
        }
      }
    },
    "/ui/loan/search/party/{partyId}": {
      "get": {
        "tags": [
          "loan-controller"
        ],
        "operationId": "getLoansByParty",
        "parameters": [
          {
            "name": "partyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMetadataLoan"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "state": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
      "ModelSpec": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Party": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/PartyContact"
          },
          "identity": {
            "$ref": "#/components/schemas/PartyIdentity"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "PartyAddress": {
        "type": "object",
        "properties": {
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "PartyContact": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/PartyAddress"
          }
        }
      },
      "PartyIdentity": {
        "type": "object",
        "properties": {
          "taxId": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "identityType": {
            "type": "string"
          },
          "identityNumber": {
            "type": "string"
          }
        }
      },
      "EntityWithMetadataParty": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Party"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
      "Loan": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "string"
          },
          "partyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "apr": {
            "type": "number"
          },
          "termMonths": {
            "type": "integer",
            "format": "int32"
          },
          "dayCountBasis": {
            "type": "string"
          },
          "repaymentDay": {
            "type": "integer",
            "format": "int32"
          },
          "fundedDate": {
            "type": "string",
            "format": "date"
          },
          "fundedAmount": {
            "type": "number"
          },
          "outstandingPrincipal": {
            "type": "number"
          },
          "accruedInterest": {
            "type": "number"
          },
          "totalInterestReceivable": {
            "type": "number"
          },
          "maturityDate": {
            "type": "string",
            "format": "date"
          },
          "nextDueDate": {
            "type": "string",
            "format": "date"
          },
          "referenceSchedule": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanScheduleItem"
            }
          },
          "approval": {
            "$ref": "#/components/schemas/LoanApproval"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "LoanApproval": {
        "type": "object",
        "properties": {
          "approvedBy": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approverRole": {
            "type": "string"
          },
          "approvalComments": {
            "type": "string"
          },
          "checkedBy": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "checkerRole": {
            "type": "string"
          }
        }
      },
      "LoanScheduleItem": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "remainingBalance": {
            "type": "number"
          }
        }
      },
      "EntityWithMetadataLoan": {
        "type": "object",
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/Loan"
          },
          "meta": {
            "$ref": "#/components/schemas/EntityMetadata"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string"
          },
          "modelKey": {
            "$ref": "#/components/schemas/ModelSpec"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "transitionForLatestSave": {
            "type": "string"
          }
        }
      },
...
    }
  }
}